a bunch of refactor, now it works
